userprof fin

+++ src/controllers/videoController.js
@@ -1,3 +1,4 @@
+import User from "../models/User";
 import Video from "../models/Video";
 
 export const home = async (req, res) => {
@@ -9,7 +10,7 @@ export const search = async (req, res) => {
   const { keyword } = req.query;
   let videos = [];
 
-  if (keyword) {
+  if (keyword !== undefined) {
     videos = await Video.find({
       title: {
         $regex: new RegExp(keyword, "i"),
@@ -22,31 +23,42 @@ export const search = async (req, res) => {
 
 export const watch = async (req, res) => {
   const { id } = req.params;
-  const video = await Video.findById(id);
+  const video = await Video.findById(id).populate("owner");
 
   if (video === null) {
     return res.status(404).render("404", { pageTitle: "Video not found" });
   }
 
-  res.render("watch", { pageTitle: video.title, video });
+  res.render("videos/watch", { pageTitle: video.title, video });
 };
 
 export const getUpload = (req, res) => {
-  res.render("upload", { pageTitle: "Upload Video" });
+  res.render("videos/upload", { pageTitle: "Upload Video" });
 };
 
 export const postUpload = async (req, res) => {
-  const { title, description, hashtags } = req.body;
+  const {
+    body: { title, description, hashtags },
+    session: { user: _id },
+    file: { path: fileUrl },
+  } = req;
 
   try {
-    await Video.create({
+    const newVideo = await Video.create({
       title,
+      fileUrl,
       description,
       hashtags: Video.formatHashtags(hashtags),
+      owner: _id,
     });
+
+    const user = await User.findById(_id);
+    user.videos.push(newVideo._id);
+    user.save();
+
     res.redirect("/");
   } catch (error) {
-    res.status(400).render("upload", {
+    res.status(400).render("videos/upload", {
       pageTitle: "Upload Video",
       errorMessage: error._message,
     });
@@ -54,23 +66,37 @@ export const postUpload = async (req, res) => {
 };
 
 export const getEdit = async (req, res) => {
-  const { id } = req.params;
+  const {
+    params: { id },
+    session: { user },
+  } = req;
   const video = await Video.findById(id);
 
   if (video === null) {
     return res.status(404).render("404", { pageTitle: "Video not found" });
   }
 
-  res.render("edit", { pageTitle: `Edit: ${video.title}`, video });
+  if (video.owner.toString() !== user._id) {
+    return res.status(403).redirect("/");
+  }
+
+  res.render("videos/edit", { pageTitle: `Edit: ${video.title}`, video });
 };
 
 export const postEdit = async (req, res) => {
-  const { id } = req.params;
-  const { title, description, hashtags } = req.body;
+  const {
+    params: { id },
+    body: { title, description, hashtags },
+    session: { user },
+  } = req;
   const video = await Video.exists({ _id: id });
 
-  if (!video) {
+  if (video === null) {
     return res.status(404).render("404", { pageTitle: "Video not found" });
+  }
+
+  if (video.owner.toString() !== user._id) {
+    return res.status(403).redirect("/");
   }
 
   await Video.findByIdAndUpdate(id, {
@@ -83,7 +109,21 @@ export const postEdit = async (req, res) => {
 };
 
 export const deleteVideo = async (req, res) => {
-  const { id } = req.params;
+  const {
+    params: { id },
+    session: { user },
+  } = req;
+
+  const video = await Video.findById(id);
+
+  if (video === null) {
+    return res.status(404).render("404", { pageTitle: "Video not found" });
+  }
+
+  if (video.owner.toString() !== user._id) {
+    return res.status(403).redirect("/");
+  }
+
   await Video.findByIdAndDelete(id);
   res.redirect("/");
 };
